Move from using before_* callbacks to prepare_changes
Callbacks have been deprecated in Ecto 1.1 (and removed in Ecto 2.0) and
they involve considerable magic, so this seemed like a good solution.

We use the `Ecto.Changeset.prepare_changes/2` function to do the
reordering inside the same transaction as the changes to the model.

This does mean that the interface must become a bit more explicit: A
call to `EctoOrdered.set_order/3` should be included in the changeset
methods. Also this means that one cannot simply delete records without
using a changeset - the ordering would be messed up if this were done.

A delete function like this is required:

```elixir
def delete(model) do
  model
  |> cast(%{}, ~w(), ~w())
  |> Map.put(:action, :delete)
  |> set_order(:position)
end
```